+++ test/unit/grammar-error.spec.js
@@ -138,6 +138,17 @@ error: message
           [],
         ]);
 
+        it("null source text", () => {
+          expect(e.format([{ source: null, text: null }])).to.equal(`\
+error: message
+ at foo.peggy:1:1
+ at foo.peggy:1:6: Subinfo
+
+warning: Warning message
+ at foo.peggy:1:6
+ at foo.peggy:1:6: Warning Subinfo`);
+        });
+
         it("with source", () => {
           expect(e.format([source])).to.equal(`\
 error: message